リファクタ

+++ src/driver/speaker.h
@@ -12,5 +12,5 @@ namespace driver::speaker
 
   void init();
   void play(note_t *notes, int size);
-  void tone(uint32_t freq, int64_t ms);
+  void tone(uint32_t freq, int32_t ms);
 }
